requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
gspread~=5.12.4
aiogram~=3.3.0
python-dotenv~=1.0.0